pyproject.toml
[project]
name = "yaml-form-engine"
version = "0.1.0"
description = "A generic YAML-driven Streamlit form engine. Define forms in YAML, render them instantly. AI generates the config, humans fill in the data — all locally."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "David Oliver" }
]

dependencies = [
    "streamlit>=1.30.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
yfe = "yaml_form_engine.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["yaml_form_engine*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
